+++ src/atoms/RouteButton.jsx
@@ -13,6 +13,7 @@ const RouteButton = (props) => {
                <>
                <StyledButton key="1" 
                onClick={props.onClick} 
+               type={props.btnType}
                style={{marginLeft:'10px' }}
                 >{props.title}
                </StyledButton>
